adicionado novas libs
- adicionado sweetalert e trocando tela de admin

+++ src/pages/PostsArea/PostsArea.jsx
@@ -7,7 +7,8 @@ import { get } from 'lodash';
 import { useSelector } from 'react-redux';
 import { useEffect, useState } from 'react';
 import Spinner from '@/components/Spinners/Spinner.jsx';
-
+import Swal from 'sweetalert2';
+import withReactContent from "sweetalert2-react-content";
 export default function PostsArea() {
   const [materias, setMaterias] = useState([]);
   const [showModal, setShowModal] = useState(false);
@@ -18,7 +19,7 @@ export default function PostsArea() {
   const [showEditModal, setShowEditModal] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
   const token = useSelector((state) => state.auth.token);
-
+  const mySwal = withReactContent(Swal)
 
   useEffect(() => {
     async function fetchMaterias() {
@@ -33,7 +34,11 @@ export default function PostsArea() {
         setMaterias(response.data);
       } catch (err) {
         setIsLoading(false);
-        toast.error('Erro ao carregar matérias.');
+        mySwal.fire({
+          title: 'Erro ao carregar as máterias.',
+          text: err,
+          icon: 'error'
+        });
 
       } finally {
         setIsLoading(false);
@@ -90,16 +95,31 @@ export default function PostsArea() {
   }
 
   const handleDeleteMateria = async (index, id) => {
-    try {
-      await axios.delete(`http://localhost:3000/materias/${id}`, {
-        headers: {
-          'Authorization': `Bearer ${token}`,
-        },
-      });
-      setMaterias((prevMaterias) => prevMaterias.filter((_, i) => i !== index));
-      toast.success('Matéria excluída com sucesso!');
-    } catch (err) {
-      toast.error('Erro ao excluir matéria.');
+    const result = await mySwal.fire({
+      title: "Você tem certeza?",
+      text: "Deseja mesmo excluir a matéria?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#820AD1',
+      confirmButtonText: 'Sim, desejo excluir!'
+    });
+
+    if (result.isConfirmed) {
+      try {
+        await axios.delete(`http://localhost:3000/materias/${id}`, {
+          headers: {
+            'Authorization': `Bearer ${token}`,
+          },
+        });
+        setMaterias((prevMaterias) => prevMaterias.filter((_, i) => i !== index));
+        mySwal.fire({
+          title: 'Excluido!',
+          text: 'A matéria foi excluida com sucesso.',
+          icon: 'success'
+        });
+      } catch (err) {
+        toast.error('Erro ao excluir matéria.');
+      }
     }
   };
 
@@ -113,7 +133,6 @@ export default function PostsArea() {
     const updatedMateria = {
       nome: editNome,
     };
-
     try {
       const response = await axios.put(`http://localhost:3000/materias/${materias[editingIndex].id}`, updatedMateria);
       const updatedMaterias = [...materias];
@@ -121,9 +140,17 @@ export default function PostsArea() {
       setMaterias(updatedMaterias);
       setEditingIndex(null);
       setShowEditModal(false);
-      toast.success('Matéria editada com sucesso!');
+      mySwal.fire({
+        title: 'Editado!',
+        text: 'A máteria foi editada com sucesso.',
+        icon: 'success'
+      });
     } catch (err) {
-      toast.error('Erro ao editar matéria.');
+      mySwal.fire({
+        title: 'Erro ao editar matéria',
+        text: err,
+        icon: 'error'
+      });
     }
   };
 
@@ -141,115 +168,60 @@ export default function PostsArea() {
             >
               Adicionar Matéria
             </button>
-            {/*{materias.length > 0 ? (*/}
-            {/*    <button*/}
-            {/*        className="border border-gray-300 px-16 xl:px-24 max-lg:w-full py-2 rounded-lg mt-4 dark:text-white"*/}
-            {/*        onClick={() => setShowModal(true)}*/}
-            {/*    >*/}
-            {/*      Adicionar Matéria*/}
-            {/*    </button>*/}
-            {/*) : (*/}
-            {/*    <button*/}
-            {/*        className="hidden border border-gray-300 px-16 xl:px-24 max-lg:w-full py-2 rounded-lg mt-4 dark:text-white "*/}
-            {/*        onClick={() => setShowModal(true)}*/}
-            {/*    >*/}
-            {/*      Adicionar Matéria*/}
-            {/*    </button>*/}
-            {/*)}*/}
           </section>
-
-
-          <section className="grid grid-cols-1 md:grid-cols-2  xl:grid-cols-3 3xl:grid-cols-4 gap-x-8 gap-y-16 justify-center mx-auto mt-8 ">
-            {materias.map((materia, index) => (
-              <div key={index} className="relative">
-                <Link to={`/dashboard/postagens/${materia.nome.toLowerCase().replace(' ', '')}`}>
-                  <SubjectCard
-                    banner={bannerMateria}
-                    nome={materia.nome}
-                    atividades={materia.atividades}
-                    subject={materia}
-                  />
-                </Link>
-                <div className="absolute top-2 right-2 p-2 rounded-full focus:outline-none z-50">
-                  <button
-                    onClick={() => setMenuVisible(menuVisible === index ? null : index)}
-                    className="px-3 py-2 rounded-full bg-gray-100 dark:bg-zinc-800"
-                    style={{ color: '#000' }}
-                  >
-                    <svg xmlns="http://www.w3.org/2000/svg" height="14" width="3.5" viewBox="0 0 128 512">
-                      <path className='fill-purplePrimary' d="M64 360a56 56 0 1 0 0 112 56 56 0 1 0 0-112zm0-160a56 56 0 1 0 0 112 56 56 0 1 0 0-112zM120 96A56 56 0 1 0 8 96a56 56 0 1 0 112 0z" /></svg>
-                  </button>
-                  {menuVisible === index && (
-                    <div
-                      className="absolute top-10 right-2 w-24 bg-white border border-gray-200 dark:border-zinc-600 rounded shadow-lg z-50">
-                      <button
-                        onClick={() => handleEditMateria(index)}
-                        className="block w-full text-left px-4 py-2 text-gray-800 hover:bg-gray-100 dark:hover:bg-zinc-900 dark:text-zinc-100"
-                      >
-                        Editar
-                      </button>
-                      <button
-                        onClick={() => handleDeleteMateria(index, materia.id)}
-                        className="block w-full text-left px-4 py-2 text-gray-800 hover:bg-gray-100 dark:text-zinc-100 dark:hover:bg-zinc-900"
-                      >
-                        Excluir
-                      </button>
-                    </div>
-                  )}
+          {materias.length > 0 ? (
+            <section className="grid grid-cols-1 md:grid-cols-2  xl:grid-cols-3 3xl:grid-cols-4 gap-x-8 gap-y-16 justify-center mx-auto mt-8 ">
+              {materias.map((materia, index) => (
+                <div key={index} className="relative">
+                  <Link to={`/dashboard/postagens/${materia.nome.toLowerCase().replace(' ', '')}`}>
+                    <SubjectCard
+                      banner={bannerMateria}
+                      nome={materia.nome}
+                      atividades={materia.atividades}
+                      subject={materia}
+                    />
+                  </Link>
+                  <div className="absolute top-2 right-2 p-2 rounded-full focus:outline-none z-50">
+                    <button
+                      onClick={() => setMenuVisible(menuVisible === index ? null : index)}
+                      className="px-3 py-2 rounded-full bg-gray-100 dark:bg-zinc-800"
+                      style={{ color: '#000' }}
+                    >
+                      <svg xmlns="http://www.w3.org/2000/svg" height="14" width="3.5" viewBox="0 0 128 512">
+                        <path className='fill-purplePrimary' d="M64 360a56 56 0 1 0 0 112 56 56 0 1 0 0-112zm0-160a56 56 0 1 0 0 112 56 56 0 1 0 0-112zM120 96A56 56 0 1 0 8 96a56 56 0 1 0 112 0z" /></svg>
+                    </button>
+                    {menuVisible === index && (
+                      <div
+                        className="absolute top-10 right-2 w-24 bg-white border border-gray-200 dark:border-zinc-600 rounded shadow-lg z-50">
+                        <button
+                          onClick={() => handleEditMateria(index)}
+                          className="block w-full text-left px-4 py-2 text-gray-800 hover:bg-gray-100 dark:hover:bg-zinc-900 dark:text-zinc-100"
+                        >
+                          Editar
+                        </button>
+                        <button
+                          onClick={() => handleDeleteMateria(index, materia.id)}
+                          className="block w-full text-left px-4 py-2 text-gray-800 hover:bg-gray-100 dark:text-zinc-100 dark:hover:bg-zinc-900"
+                        >
+                          Excluir
+                        </button>
+                      </div>
+                    )}
+                  </div>
                 </div>
-              </div>
-            ))}
-            {/*{materias.length > 0 ? (*/}
-            {/*    materias.map((materia, index) => (*/}
-            {/*        <div key={index} className="relative">*/}
-            {/*          <Link to={`/dashboard/postagens/${materia.nome.toLowerCase().replace(' ', '')}`}>*/}
-            {/*            <SubjectCard*/}
-            {/*                banner={bannerMateria}*/}
-            {/*                nome={materia.nome}*/}
-            {/*                atividades={materia.atividades}*/}
-            {/*                subject={materia}*/}
-            {/*            />*/}
-            {/*          </Link>*/}
-            {/*          <div className="absolute top-2 right-2 p-2 rounded-full focus:outline-none z-50">*/}
-            {/*            <button*/}
-            {/*                onClick={() => setMenuVisible(menuVisible === index ? null : index)}*/}
-            {/*                className="p-2 rounded-full bg-gray-200 dark:bg-white"*/}
-            {/*                style={{ color: '#000' }}*/}
-            {/*            >*/}
-            {/*              &#x22EE;*/}
-            {/*            </button>*/}
-            {/*            {menuVisible === index && (*/}
-            {/*                <div*/}
-            {/*                    className="absolute top-10 right-2 w-24 bg-white border border-gray-200 dark:border-zinc-600 rounded-lg shadow-lg z-50">*/}
-            {/*                  <button*/}
-            {/*                      onClick={() => handleEditMateria(index)}*/}
-            {/*                      className="block w-full text-left px-4 py-2 text-gray-800 hover:bg-gray-100 dark:hover:bg-zinc-900 dark:text-zinc-100"*/}
-            {/*                  >*/}
-            {/*                    Editar*/}
-            {/*                  </button>*/}
-            {/*                  <button*/}
-            {/*                      onClick={() => handleDeleteMateria(index, materia.id)}*/}
-            {/*                      className="block w-full text-left px-4 py-2 text-gray-800 hover:bg-gray-100 dark:text-zinc-100 dark:hover:bg-zinc-900"*/}
-            {/*                  >*/}
-            {/*                    Excluir*/}
-            {/*                  </button>*/}
-            {/*                </div>*/}
-            {/*            )}*/}
-            {/*          </div>*/}
-            {/*        </div>*/}
-            {/*    ))*/}
-            {/*) : (*/}
-            {/*    <p className="col-span-4 text-center text-gray-500 dark:text-zinc-200">*/}
-            {/*      Não foi possivel encontrar as matérias.*/}
-            {/*    </p>*/}
-            {/*)}*/}
-          </section>
+              ))}
+            </section>
+          ) : (
+            <p className="text-center text-gray-600 dark:text-gray-300 mt-8 text-xl">
+              Não foi possível encontrar as matérias.
+            </p>
+          )}
         </main>
       </div>
 
       {showModal && (
         <div
-          className="fixed inset-0 flex items-center justify-center bg-gray-800 bg-opacity-50 "
+          className="fixed inset-0 flex items-center justify-center bg-gray-800 bg-opacity-50 z-50"
           onClick={() => setShowModal(false)}>
 
           <div
@@ -280,7 +252,7 @@ export default function PostsArea() {
 
       {showEditModal && (
         <div
-          className="fixed inset-0 flex items-center justify-center bg-gray-800 bg-opacity-50  max-lg:px-4"
+          className="fixed inset-0 flex items-center justify-center bg-gray-800 bg-opacity-50  max-lg:px-4 z-50"
           onClick={() => setShowEditModal(false)}>
           <div
             className="bg-white py-8 lg:w-1/3 md:w-1/2 w-full lg:px-16 px-0 rounded-lg shadow-lg lg:text-left max-lg:px-4"
